olympus/src/main/java/com/olympus/utils/AppUtils.java
package com.olympus.utils;

import com.olympus.dto.response.report.WeeklyReportData;
import org.apache.poi.ss.usermodel.*;
import org.apache.poi.xssf.usermodel.XSSFFont;
import org.apache.poi.xssf.usermodel.XSSFWorkbook;

import java.util.Random;

public class AppUtils {
    public static String generateRandomOTP() {
        Random random = new Random();
        int number = random.nextInt(999999);
        return String.format("%06d", number);
    }

    public static Workbook generateWeeklyReport(WeeklyReportData data) {
        XSSFWorkbook workbook = new XSSFWorkbook();

        Sheet sheet = workbook.createSheet("Weekly Report");
        sheet.setColumnWidth(0, 6000);
        sheet.setColumnWidth(1, 6000);
        sheet.setColumnWidth(2, 6000);
        sheet.setColumnWidth(3, 6000);

        Row header = sheet.createRow(0);

        CellStyle headerStyle = workbook.createCellStyle();
        headerStyle.setFillForegroundColor(IndexedColors.LIGHT_BLUE.getIndex());
        headerStyle.setFillPattern(FillPatternType.SOLID_FOREGROUND);

        XSSFFont font = workbook.createFont();
        font.setFontName("Arial");
        font.setFontHeightInPoints((short) 16);
        font.setBold(true);
        headerStyle.setFont(font);

        Cell headerCell = header.createCell(0);
        headerCell.setCellValue("New Post");
        headerCell.setCellStyle(headerStyle);

        headerCell = header.createCell(1);
        headerCell.setCellValue("New Friends");
        headerCell.setCellStyle(headerStyle);

        headerCell = header.createCell(2);
        headerCell.setCellValue("New Likes");
        headerCell.setCellStyle(headerStyle);

        headerCell = header.createCell(3);
        headerCell.setCellValue("New Comments");
        headerCell.setCellStyle(headerStyle);

        CellStyle style = workbook.createCellStyle();
        style.setWrapText(true);

        Row row = sheet.createRow(1);
        Cell cell = row.createCell(0);
        cell.setCellValue(data.getNewPostsLastWeek());
        cell.setCellStyle(style);

        cell = row.createCell(1);
        cell.setCellValue(data.getNewFriendsLastWeek());
        cell.setCellStyle(style);

        cell = row.createCell(2);
        cell.setCellValue(data.getNewLikesLastWeek());
        cell.setCellStyle(style);

        cell = row.createCell(3);
        cell.setCellValue(data.getNewCommentsLastWeek());
        cell.setCellStyle(style);

        return workbook;
    }
}
